Code Cleanup
Made code more readable, as well as some optimizations removing redundant code. the EconomyManager has a new method for creating the balance.dat file. As there were issues when the file just wouldn't create. balance.dat now has a recovery mechanism if it fails it will retry 3 times to create the file. The ItemManager has been updated to pass a NamespacedKey for a Persistent Data Container, so in the future items are a lot easier to distinguish from one another.